\name{hyper.fit package}
\alias{hyper.fit-package}
\docType{package}
\title{
\packageTitle{hyper.fit}
}
\description{
\packageDescription{hyper.fit}
}
\details{
\tabular{ll}{
Package: \tab hyper.fit\cr
Type: \tab Package\cr
Version: \tab 1.1.1\cr
Date: \tab 2019-12-05\cr
License: \tab GPL-3\cr
}
Most users will interact with this package through the high level hyper.fit and hyper.plot functions. The utility functions to construct covariance matrices and arrays will also be useful (see makecovmat2d and makecovarray2d).
}
\author{
Aaron Robotham and Danail Obreschkow

Maintainer: Aaron Robotham <aaron.robotham@uwa.edu.au>
}
\references{
Robotham, A.S.G., & Obreschkow, D., PASA, 2015, 32, 33
}
\seealso{
  \code{\link{hyper.basic}}, \code{\link{hyper.convert}}, \code{\link{hyper.data}}, \code{\link{hyper.fit}}, \code{\link{hyper.plot}}, \code{\link{hyper.sigcor}}, \code{\link{hyper.summary}}
}
\keyword{ package }
\keyword{ fit }
\keyword{ hyper }
\keyword{ linear }
\keyword{ plane }
\keyword{ regression }

